Updating the directory structure for the RBUploader component.
