+++ resources/assets/components/ReportbackUploader/ReportbackUploaderContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { get } from 'lodash';
-import ReportbackUploader from '../components/ReportbackUploader';
-import { submitReportback, addToSubmissionsList, fetchUserReportbacks } from '../actions';
+import ReportbackUploader from './ReportbackUploader';
+import { submitReportback, addToSubmissionsList, fetchUserReportbacks } from '../../actions';
 
 /**
  * Provide state from the Redux store as props for this component.
@@ -12,6 +12,7 @@ const mapStateToProps = state => ({
   submissions: state.submissions,
   noun: get(state.campaign.additionalContent, 'noun'),
   userId: state.user.id,
+  uploads: state.uploads,
 });
 
 /**
