+++ resources/assets/reducers/index.js
@@ -1,12 +1,13 @@
-export campaign from './campaign';
-export reportbacks from './reportbacks';
-export submissions from './submissions';
 export blocks from './blocks';
-export user from './user';
-export signups from './signups';
-export share from './share';
-export events from './events';
-export notifications from './notifications';
+export campaign from './campaign';
 export competitions from './competitions';
+export events from './events';
 export experiments from './experiments';
 export modal from './modal';
+export notifications from './notifications';
+export reportbacks from './reportbacks';
+export share from './share';
+export signups from './signups';
+export submissions from './submissions';
+export uploads from './uploads';
+export user from './user';
