+++ resources/assets/reducers/uploads.js
@@ -1,0 +1,8 @@
+/**
+ * Uploads reducer.
+ */
+const uploads = (state = {}) => (
+  state
+);
+
+export default uploads;
